// web-rally/src/client/types.gen.ts
export type CheckPointCreate = {
    /**
     * Name
     */
    name: string;
    /**
     * Description
     */
    description: string;
    /**
     * Latitude
     */
    latitude?: number | null;
    /**
     * Longitude
     */
    longitude?: number | null;
    /**
     * Order
     */
    order: number;
};

/**
 * CheckPointUpdate
 */
export type CheckPointUpdate = {
    /**
     * Name
     */
    name?: string | null;
    /**
     * Description
     */
    description?: string | null;
    /**
     * Latitude
     */
    latitude?: number | null;
    /**
     * Longitude
     */
    longitude?: number | null;
    /**
     * Order
     */
    order?: number | null;
};

/**
 * CheckpointAssignmentUpdate
 */
export type CheckpointAssignmentUpdate = {
    /**
     * Checkpoint Id
     */
    checkpoint_id?: number | null;
};

/**
 * DetailedCheckPoint
 */
export type DetailedCheckPoint = {
    /**
     * Name
     */
    name: string;
    /**
     * Description
     */
    description: string;
    /**
     * Latitude
     */
    latitude?: number | null;
    /**
     * Longitude
     */
    longitude?: number | null;
    /**
     * Order
     */
    order: number;
    /**
     * Id
     */
    id: number;
};