// web-rally/src/client/models/CheckPointCreate.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type CheckPointCreate = {
    name: string;
    description?: (string | null);
    latitude?: (number | null);
    longitude?: (number | null);
    order: number;
};