// web-rally/src/client/models/CheckPointUpdate.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type CheckPointUpdate = {
    name?: (string | null);
    description?: (string | null);
    latitude?: (number | null);
    longitude?: (number | null);
    order?: (number | null);
};